--- runs/YOLO11/app.py
@@ -1,44 +1,0 @@
-import gradio as gr
-import PIL.Image as Image
-
-from ultralytics import ASSETS, YOLO
-
-model = YOLO("yolo11n.pt")
-
-
-def predict_image(img, conf_threshold, iou_threshold):
-    """Predicts objects in an image using a YOLO11 model with adjustable confidence and IOU thresholds."""
-    results = model.predict(
-        source=img,
-        conf=conf_threshold,
-        iou=iou_threshold,
-        show_labels=True,
-        show_conf=True,
-        imgsz=640,
-    )
-
-    for r in results:
-        im_array = r.plot()
-        im = Image.fromarray(im_array[..., ::-1])
-
-    return im
-
-
-iface = gr.Interface(
-    fn=predict_image,
-    inputs=[
-        gr.Image(type="pil", label="Upload Image"),
-        gr.Slider(minimum=0, maximum=1, value=0.25, label="Confidence threshold"),
-        gr.Slider(minimum=0, maximum=1, value=0.45, label="IoU threshold"),
-    ],
-    outputs=gr.Image(type="pil", label="Result"),
-    title="Ultralytics Gradio",
-    description="Upload images for inference. The Ultralytics YOLO11n model is used by default.",
-    examples=[
-        [ASSETS / "bus.jpg", 0.25, 0.45],
-        [ASSETS / "zidane.jpg", 0.25, 0.45],
-    ],
-)
-
-if __name__ == "__main__":
-    iface.launch(server_name="0.0.0.0")
