+++ runs/YOLO11/predict/README.md
@@ -46,3 +46,16 @@ To deploy the YOLO11 Gradio app, follow the steps below:
     ```
 
 For additional information and support, please refer to the [VESSL documentation](https://docs.vessl.ai).
+
+## Citation
+```bibtex
+@software{yolo11_ultralytics,
+  author = {Glenn Jocher and Jing Qiu},
+  title = {Ultralytics YOLO11},
+  version = {11.0.0},
+  year = {2024},
+  url = {https://github.com/ultralytics/ultralytics},
+  orcid = {0000-0001-5950-6979, 0000-0002-7603-6750, 0000-0003-3783-7069},
+  license = {AGPL-3.0}
+}
+```
